preparation for dual gt/nx support
