+++ tests/test_molecule.py
@@ -7,7 +7,7 @@ import numpy as np
 import pytest
 import qcelemental as qcel
 
-from spyrmsd import io, molecule, utils
+from spyrmsd import io, molecule, utils, graph
 from tests import molecules
 
 
@@ -164,11 +164,11 @@ def test_graph_from_adjacency_matrix(mol: molecule.Molecule, n_bonds: int) -> No
 
     G = mol.to_graph()
 
-    assert G.num_vertices() == len(mol)
-    assert G.num_edges() == n_bonds
+    assert graph.num_vertices(G) == len(mol)
+    assert graph.num_edges(G) == n_bonds
 
     for idx, atomicnum in enumerate(mol.atomicnums):
-        assert G.vertex_properties["atomicnum"][idx] == atomicnum
+        assert graph.vertex_property(G, "atomicnum", idx) == atomicnum
 
 
 @pytest.mark.parametrize(
@@ -189,8 +189,8 @@ def test_graph_from_atomic_coordinates_perception(
         # Uses automatic bond perception
         G = m.to_graph()
 
-        assert G.num_vertices() == len(m)
-        assert G.num_edges() == n_bonds
+        assert graph.num_vertices(G) == len(m)
+        assert graph.num_edges(G) == n_bonds
 
         for idx, atomicnum in enumerate(mol.atomicnums):
-            assert G.vertex_properties["atomicnum"][idx] == atomicnum
+            assert graph.vertex_property(G, "atomicnum", idx) == atomicnum
